chore: rename datasetgenerator to make_dataset
