+++ documents/passport_bottom.py
@@ -4,11 +4,11 @@ from faker import Faker
 from PIL import Image
 from PIL.Image import Resampling
 
-from datasetgenerator import DocumentGenerator
-from datasetgenerator.factories.fields import FieldFactory
-from datasetgenerator.fields import FieldGroup
-from datasetgenerator.positioning import Point
-from datasetgenerator.rendering import Font, renderers
+from make_dataset import DocumentGenerator
+from make_dataset.factories.fields import FieldFactory
+from make_dataset.fields import FieldGroup
+from make_dataset.positioning import Point
+from make_dataset.rendering import Font, renderers
 
 faker = Faker('ru_RU')
