+++ documents/passport_top.py
@@ -4,11 +4,11 @@ from textwrap import fill
 from faker import Faker
 from PIL import Image
 
-from datasetgenerator import DocumentGenerator
-from datasetgenerator.factories.fields import FieldFactory
-from datasetgenerator.fields import FieldGroup
-from datasetgenerator.positioning import Point
-from datasetgenerator.rendering import renderers
+from make_dataset import DocumentGenerator
+from make_dataset.factories.fields import FieldFactory
+from make_dataset.fields import FieldGroup
+from make_dataset.positioning import Point
+from make_dataset.rendering import renderers
 from documents.passport_bottom import field_font, series_renderer
 
 _department_names = [
