+++ make_dataset/rendering/renderers.py
@@ -4,8 +4,8 @@ from PIL import ImageDraw
 from PIL.Image import Image, Resampling, alpha_composite
 from PIL.Image import new as new_image
 
-from datasetgenerator.fields import Field
-from datasetgenerator.rendering import Font, Renderer
+from make_dataset.fields import Field
+from make_dataset.rendering import Font, Renderer
 
 
 class OpacityRenderer(Renderer):
